test: fold EPUB keyword checks into weread_tag_spec
